requirements.txt
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.4.0
Pillow>=8.2.0 